ser/de fix for CPU

+++ nd4j/nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/shape/Shape.java
@@ -1327,7 +1327,7 @@ public class Shape {
         if (rank > 2 || rank < 1)
             return false;
         else {
-            int len = Shape.length(shapeInfo);
+            long len = Shape.length(shapeInfo);
             DataBuffer shape = Shape.shapeOf(shapeInfo);
             return shape.getInt(0) == len || shape.getInt(1) == len;
         }
@@ -2476,13 +2476,14 @@ public class Shape {
      * @param buffer the buffer to get the rank for
      * @return the rank for the shape buffer
      */
-    public static int length(DataBuffer buffer) {
-        int ret = 1;
+    public static long length(DataBuffer buffer) {
+        long ret = 1;
         val rr = buffer.asLong();
         DataBuffer shape = Shape.shapeOf(buffer);
         int rank = Shape.rank(buffer);
         for (int i = 0; i < rank; i++)
             ret *= shape.getLong(i);
+
         return ret;
     }
